apply logic in different file
